server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich[toml]>=0.8.1",
    "bcrypt>=4.2.1",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.4.0",
    "tortoise-orm>=0.24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.aerich]
tortoise_orm = "src.config.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
